Fix delete workspace

+++ src/services/workspace.service.js
@@ -64,7 +64,7 @@ class WorkspaceService {
     static async updateById(workspace_id, update_workspace){
         try {
             await WorkSpaceRepository.updateById(workspace_id, update_workspace)
-            const workspace_updated = WorkSpaceRepository.getById(workspace_id)
+            const workspace_updated = await WorkSpaceRepository.getById(workspace_id)
             return workspace_updated
         } catch (error) {
             console.error(error.message)
@@ -74,13 +74,18 @@ class WorkspaceService {
 
     static async deleteById(workspace_id){
         try {
+            // Obtener el workspace antes de eliminarlo
+            const workspace_to_delete = await WorkSpaceRepository.getById(workspace_id)
+            if (!workspace_to_delete) {
+                throw new ServerError(404, 'Espacio de trabajo no encontrado')
+            }
+            // Eliminar el workspace
             await WorkSpaceRepository.deleteById(workspace_id)
-            const workspace_deleted = WorkSpaceRepository.getById(workspace_id)
-            return workspace_deleted
-            
+            // Devolver el workspace que fue eliminado
+            return workspace_to_delete
         } catch (error) {
             console.error(error.message)
-            throw new ServerError('Error interno')
+            throw error
         }
     }
 }
